requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
pypdf>=4.2.0
